requirements.txt
google-generativeai>=0.8.0
flask>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
tweepy>=4.14.0
linkedin-api>=2.0.0
google-adk>=0.5.0